# public/swagger-ui/swagger.yaml
openapi: 3.0.0
info:
  title: 'Vide grenier en ligne API'
  description: 'Une API pour gérer les articles et les villes du vide grenier en ligne'
  version: 1.0.0
paths:
  /api/products:
    get:
      summary: "Affiche la liste des articles / produits pour la page d'accueil"
      operationId: e35803db7dc3263e3622538255f4c890
      parameters:
        -
          name: sort
          in: query
          description: 'Critère de tri des produits (views ou date)'
          required: false
          schema:
            type: string
            enum:
              - views
              - date
        -
          name: latitude
          in: query
          description: 'Latitude pour la géolocalisation'
          required: false
          schema:
            type: number
            format: float
        -
          name: longitude
          in: query
          description: 'Longitude pour la géolocalisation'
          required: false
          schema:
            type: number
            format: float
        -
          name: radius
          in: query
          description: 'Rayon en kilomètres pour la géolocalisation'
          required: false
          schema:
            type: number
            format: float
        -
          name: search
          in: query
          description: 'Recherche dans les articles'
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 'Fetches products'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id: { type: integer }, name: { type: string }, description: { type: string }, published_date: { type: string }, user_id: { type: integer }, views: { type: integer }, picture: { type: string }, ville_id: { type: integer } }
                  type: object
  /api/userproducts:
    get:
      summary: 'Affiche la liste des articles / produits pour un utilisateur'
      operationId: d0cf98fe01d84a4748d9fc5a59ff48a0
      parameters:
        -
          name: sort
          in: query
          description: 'Critère de tri des produits (views ou date)'
          required: false
          schema:
            type: string
            enum:
              - views
              - date
      responses:
        '200':
          description: 'Fetches user products'
          content:
            application/json:
              schema:
                type: array
                items:
                  properties: { id: { type: integer }, name: { type: string }, description: { type: string }, published_date: { type: string }, user_id: { type: integer }, views: { type: integer }, picture: { type: string }, ville_id: { type: integer } }
                  type: object
...